% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRupt2.R
\name{eRupt2}
\alias{eRupt2}
\title{Plots a Volcano Plot}
\usage{
eRupt2(results, alpha = 0.05, FCsig = 1, Gene = "Gene")
}
\arguments{
\item{results}{Dataframe of results}

\item{alpha}{alpha value for significance}

\item{FCsig}{Fold change value for significance}

\item{Gene}{Column for labeling genes}
}
\value{
A dataframe of Ensembl IDs, gene symbols, and any additional
attributes (if chosen)
}
\description{
This function plots a volcano plot for RNAseq results
}
