% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebGestlat_Pipe2.R
\name{WebGestalt_Pipe2}
\alias{WebGestalt_Pipe2}
\title{WebGestalt Enrichment}
\usage{
WebGestalt_Pipe2(
  mode = "ORA",
  results,
  alpha = 0.05,
  FC = 0.5,
  DB = NULL,
  species = "mmusculus",
  projectName = NULL,
  geneID = "Gene",
  outputTerms = 100
)
}
\arguments{
\item{mode}{"ORA" or "GSEA". Note: GSEA mode runs GSEA on the log2FC values.}

\item{results}{A makeComp() output data.frame. Any data frame with Gene name, pAdj, and log2FC can also work. Gene names must be in column 7, and log2foldchange must be in column 2. Columns must be named: "Gene", "padj", and "log2FoldChange".}

\item{alpha}{pAdj cut-off for DEGs (ORA only)}

\item{FC}{absolute log2FC cut-off for DEGs (ORA only)}

\item{DB}{WebGestalt enrichDatabase to query against, see WebGestaltR::listGeneSet() for a full list.}

\item{species}{WebGestalt organism, see WebGestaltR::listOrganism() for a full list.}

\item{projectName}{Name of raw data folder.}

\item{geneID}{Column for genes to use for enrichment}

\item{outputTerms}{The threshold for the top method. The default is 100.}
}
\value{
A dataframe of Ensembl IDs, gene symbols, and any additional
attributes (if chosen)
}
\description{
This function calculates enrichments results using the DEG dataframe
}
